+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - better highlight in xml doc comments
 - highlight inline code between backticks or single quotes in xml doc tooltips
+- command to reset FSI arguments to default values
 ### Changed
 - silently reload externally changed files when the editor has no unsaved changes (configurable via `AutoReloadExternalChangesIfClean` setting)
 - File menu: added "Auto-Reload External Changes" toggle to control the above